refactor: Rename clickhouse.resources.py to client.py.
